update nbs to reflect valid data updates
These are small improvements, that should have been made long ago.
They introduce minor changes to the data, but not to values that we use downstream.
The heatmap plot needs to be regenerated, otherwise the changes create no downstream liabilites.
Changes:

- Move treatment of Mon078 from code level to DB level. This ends a stupid temporal solution where Mon078 was removed in the notebook. Now the valid column of the experiments table has been updated and Mon078 is removed like all the other invalid building blocks
- Rerun add_labels_to_db for 2023-12-20 data. This introduces minor changes to the "scaled" responses and consequently the "major" label which changes for 14 out of 40k data points.
- Minor code and doc improvements

These changes address discrapancies in the number of retrieved data points when querying the DB with slightly different methods. Now all returned counts align with the expectation and the values in the SI draft.